backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
pydantic[email]>=2.10.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.25.2
user-agents==2.2.0
geoip2==4.7.0
requests==2.31.0
apscheduler==3.10.4
websockets==12.0
bcrypt==4.1.1
